+++ backend/controllers/user_controller.js
@@ -1,12 +1,131 @@
 const userModel = require(`../models/index`).user
 const Op = require(`sequelize`).Op
-const upload = require(`./upload`).single(`foto`)
+const upload = require(`./upload_user`).single(`foto`)
 const path = require(`path`)
 const fs = require(`fs`)
 const md5 = require("md5")
 
 const jsonwebtoken = require("jsonwebtoken")
 const SECRET_KEY = "secretcode"
+
+exports.login = async (request, response) => {
+    try {
+        const params = {
+            email: request.body.email,
+            password: md5(request.body.password),
+        };
+
+        const findUser = await userModel.findOne({ where: params });
+        if (findUser == null) {
+            return response.status(400).json({
+                message: "email or password doesn't match"
+            });
+        }
+        console.log(findUser);
+        //generate jwt token
+        let tokenPayLoad = {
+            id: findUser.id,
+            email: findUser.email,
+            role: findUser.role,
+            nama_user: findUser.nama_user,
+        };
+        tokenPayLoad = JSON.stringify(tokenPayLoad);
+        let token = await jsonwebtoken.sign(tokenPayLoad, SECRET_KEY);
+
+        return response.status(200).json({
+            message: "Success login",
+            data: {
+                token: token,
+                id: findUser.id,
+                nama_user: findUser.nama_user,
+                email: findUser.email,
+                role: findUser.role,
+            },
+        });
+    } catch (error) {
+        console.log(error);
+        return response.status(400).json({
+            message: error,
+        });
+    }
+}
+
+exports.register = (request, response) => {
+    upload(request, response, async error => {
+        if (error) {
+            return response.json({
+                message: error
+            })
+        }
+        if (!request.file) {
+            return response.json({
+                message: `Nothing to Upload`
+            })
+        }
+
+        let newUser = {
+            nama_user: request.body.nama_user,
+            email: request.body.email,
+            password: md5(request.body.password),
+            role: 'customer',
+            foto: request.file.filename
+        }
+
+
+        let user = await userModel.findAll({
+            where: {
+                [Op.or]: [{ nama_user: newUser.nama_user }, { email: newUser.email }],
+            },
+        });
+
+        if (
+            newUser.nama_user === "" ||
+            newUser.email === "" ||
+            newUser.password === ""
+        ) {
+            const oldFotoUser = newUser.foto;
+            const patchFoto = path.join(__dirname, `../gambar/user`, oldFotoUser);
+            if (fs.existsSync(patchFoto)) {
+                fs.unlink(patchFoto, (error) => console.log(error));
+            }
+
+            return response.status(400).json({
+                success: false,
+                message: "Harus diisi semua",
+            });
+        } else {
+            //nama dan email tidak boleh sama
+            if (user.length > 0) {
+                const oldFotoUser = newUser.foto;
+                const patchFoto = path.join(__dirname, `../gambar/user`, oldFotoUser);
+                if (fs.existsSync(patchFoto)) {
+                    fs.unlink(patchFoto, (error) => console.log(error));
+                }
+                return response.status(400).json({
+                    success: false,
+                    message: "Cari nama atau email lain",
+                });
+            } else {
+                console.log(newUser);
+                userModel
+                    .create(newUser)
+                    .then((result) => {
+                        return response.json({
+                            success: true,
+                            data: result,
+                            message: `New User has been inserted`,
+                        });
+                    })
+                    .catch((error) => {
+                        return response.status(400).json({
+                            success: false,
+                            message: error.message,
+                        });
+                    });
+            }
+        }
+    })
+}
 
 exports.getAllUser = async (request, response) => {
     let users = await userModel.findAll()
@@ -17,72 +136,15 @@ exports.getAllUser = async (request, response) => {
     })
 }
 
-exports.findUser = async (request, response) => {
-    let nama_user = request.body.nama_user
-    let users = await userModel.findAll({
-        where: {
-            [Op.or]: [
-                { nama_user: { [Op.substring]: nama_user } }
-            ]
-        }
-    })
-    return response.json({
-        success: true,
-        data: users,
-        message: `All users have been loaded`
-    })
-}
-
-exports.login = async (request, response) => {
-    try {
-        const params = {
-            email: request.body.email,
-            password: md5(request.body.password),
-        }
-
-        const findUser = await userModel.findOne({ where: params })
-        if (findUser == null) {
-            return response.status(404).json({
-                message: "email or password doesn't match",
-                err: error,
-            })
-        }
-        console.log(findUser)
-        let tokenPayload = {
-            id_user: findUser.id_customer,
-            email: findUser.email,
-            role: findUser.role,
-        }
-        tokenPayload = JSON.stringify(tokenPayload)
-        let token = await jsonwebtoken.sign(tokenPayload, SECRET_KEY)
-
-        return response.status(200).json({
-            message: "Succes Login",
-            data: {
-                token: token,
-                id_user: findUser.id_user,
-                email: findUser.email,
-                role: findUser.role,
-            }
-        })
-    } catch (error) {
-        console.log(error)
-        return response.status(500).json({
-            message: "Internal error",
-            err: error,
-        })
-    }
-}
-       
 exports.addUser = (request, response) => {
     upload(request, response, async error => {
         if (error) {
-            return response.json({ 
-                message: error 
+            return response.json({
+                message: error
             })
         }
         if (!request.file) {
-            return response.json({ 
+            return response.json({
                 message: `Nothing to Upload`
             })
         }
@@ -95,22 +157,61 @@ exports.addUser = (request, response) => {
             foto: request.file.filename
         }
 
-        
-        userModel.create(newUser)
-            .then(result => {
-                return response.json({
-                    success: true,
-                    data: result,
-                    message: `New User has been inserted`
-                })
-            })
-            .catch(error => {
-                return response.json({
+
+        let user = await userModel.findAll({
+            where: {
+                [Op.or]: [{ nama_user: newUser.nama_user }, { email: newUser.email }],
+            },
+        });
+
+        if (
+            newUser.nama_user === "" ||
+            newUser.email === "" ||
+            newUser.password === "" ||
+            newUser.role === ""
+        ) {
+            const oldFotoUser = newUser.foto;
+            const patchFoto = path.join(__dirname, `../gambar/user`, oldFotoUser);
+            if (fs.existsSync(patchFoto)) {
+                fs.unlink(patchFoto, (error) => console.log(error));
+            }
+
+            return response.status(400).json({
+                success: false,
+                message: "Harus diisi semua",
+            });
+        } else {
+            //nama dan email tidak boleh sama
+            if (user.length > 0) {
+                const oldFotoUser = newUser.foto;
+                const patchFoto = path.join(__dirname, `../gambar/user`, oldFotoUser);
+                if (fs.existsSync(patchFoto)) {
+                    fs.unlink(patchFoto, (error) => console.log(error));
+                }
+                return response.status(400).json({
                     success: false,
-                    message: error.message
-                })
-            })
-    })        
+                    message: "Cari nama atau email lain",
+                });
+            } else {
+                console.log(newUser);
+                userModel
+                    .create(newUser)
+                    .then((result) => {
+                        return response.json({
+                            success: true,
+                            data: result,
+                            message: `New User has been inserted`,
+                        });
+                    })
+                    .catch((error) => {
+                        return response.status(400).json({
+                            success: false,
+                            message: error.message,
+                        });
+                    });
+            }
+        }
+    })
 }
 
 exports.updateUser = (request, response) => {
@@ -120,6 +221,20 @@ exports.updateUser = (request, response) => {
         }
 
         let id = request.params.id
+
+        let getId = await userModel.findAll({
+            where: {
+                [Op.and]: [{ id: id }],
+            },
+        });
+
+        if (getId.length === 0) {
+            return response.status(400).json({
+                success: false,
+                message: "User dengan id tersebut tidak ada",
+            });
+        }
+
         let dataUser = {
             nama_user: request.body.nama_user,
             email: request.body.email,
@@ -132,35 +247,85 @@ exports.updateUser = (request, response) => {
                 where: { id: id }
             })
             const oldFoto = selectedFoto.foto
-            const pathFoto = path.join(__dirname, `../gambar`, oldFoto)
+            const pathFoto = path.join(__dirname, `../gambar/user`, oldFoto)
             if (fs.existsSync(pathFoto)) {
                 fs.unlink(pathFoto, error =>
-                console.log(error))
+                    console.log(error))
             }
             dataUser.foto = request.file.filename
         }
-        
-        userModel.update(dataUser, { where: { id: id } })
-            .then(result => {
-                return response.json({
-                    success: true,
-                    message: `Data user has been updated`
-                })
-            })
-            .catch(error => {
-                return response.json({
+
+        let user = await userModel.findAll({
+            where: {
+                [Op.or]: [{ nama_user: dataUser.nama_user }, { email: dataUser.email }],
+            },
+        });
+        if (
+            dataUser.nama_user === "" ||
+            dataUser.email === "" ||
+            dataUser.password === "" ||
+            dataUser.role === ""
+        ) {
+            const oldFotoUser = dataUser.foto;
+            const patchFoto = path.join(__dirname, `../gambar/user`, oldFotoUser);
+            if (fs.existsSync(patchFoto)) {
+                fs.unlink(patchFoto, (error) => console.log(error));
+            }
+
+            return response.status(400).json({
+                success: false,
+                message: "Harus diisi semua",
+            });
+        } else {
+
+            if (user.length > 0) {
+                const oldFotoUser = dataUser.foto;
+                const patchFoto = path.join(__dirname, `../gambar/user`, oldFotoUser);
+                if (fs.existsSync(patchFoto)) {
+                    fs.unlink(patchFoto, (error) => console.log(error));
+                }
+                return response.status(400).json({
                     success: false,
-                    message: error.message
+                    message: "Cari nama atau email lain",
+                });
+            }
+
+            userModel.update(dataUser, { where: { id: id } })
+                .then(result => {
+                    return response.json({
+                        success: true,
+                        message: `Data user has been updated`
+                    })
                 })
-            })
+                .catch(error => {
+                    return response.json({
+                        success: false,
+                        message: error.message
+                    })
+                })
+        }
     })
 }
 
 exports.deleteUser = async (request, response) => {
     let id = request.params.id
+
+    let getId = await userModel.findAll({
+        where: {
+            [Op.and]: [{ id: id }],
+        },
+    });
+
+    if (getId.length === 0) {
+        return response.status(400).json({
+            success: false,
+            message: "User dengan id tersebut tidak ada",
+        });
+    }
+
     const user = await userModel.findOne({ where: { id: id } })
     const oldFoto = user.foto
-    const pathFoto = path.join(__dirname, '../gambar', oldFoto)
+    const pathFoto = path.join(__dirname, '../gambar/user', oldFoto)
 
     if (fs.existsSync(pathFoto)) {
         fs.unlink(pathFoto, error => console.log(error))
